move all definitions under `definition` folder
