+++ mods/grinder/src/grinder/definition/node.lua
@@ -1,0 +1,5 @@
+
+return {
+	active = require('grinder.definition.node.active'),
+	inactive = require('grinder.definition.node.inactive'),
+}
