--- mods/grinder/src/grinder/node.lua
@@ -1,5 +1,0 @@
-
-return {
-	active = require('grinder.node.active'),
-	inactive = require('grinder.node.inactive'),
-}
